backend/requirements.txt
requests 
opencv-python
azure-ai-vision-face
python-dotenv